disabled the deleting of a category with associated posts. Disabled the deleting of tags associated with posts.

+++ cms/app/Http/Controllers/CategoriesController.php
@@ -105,6 +105,13 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count()>0) {
+
+            session()->flash('error', 'Category cannot be deleted because it has some posts');
+  
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully');
